+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [UNRELEASED]
+## [1.0.0] - 2017-01-24
 
 ### Added
 
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Upgraded to the final version of PSR-15 `psr/http-server-middleware`
+- Renamed the option `get()` to `getMethods()`
+- Renamed the option `post()` to `postMethods()`
 
 ### Fixed